Fatemas branch

+++ Group6System-main/Group6System-main/ApplicationTier/src/main/java/com/example/applicationtier/DAO/admin/AdminDAOImpl.java
@@ -43,7 +43,7 @@ public class AdminDAOImpl implements AdminDAO {
     }
 
     @Override
-    public String removeCustomer(int cprNumber){
+    public String removeCustomer(long cprNumber){
         Request obj = new Request("RemoveCustomerByCprNumber", cprNumber);
         handler.setObj(obj);
