+++ Group6System-main/Group6System-main/ApplicationTier/src/main/java/com/example/applicationtier/service/adminservice/AdminServiceImpl.java
@@ -31,7 +31,7 @@ public class AdminServiceImpl implements AdminService{
     }
 
     @Override
-    public String removeCustomer(int cprNumber){
+    public String removeCustomer(long cprNumber){
         String customerToRemove = adminDAO.removeCustomer(cprNumber);
         return customerToRemove;
     }
