+++ Group6System-main/Group6System-main/BlazorClient/BlazorClient/Data/AdminValidation/IAdminService.cs
@@ -8,7 +8,7 @@ namespace BlazorClient.Data.AdminValidation
     public interface IAdminService
     {
         Task ValidateCustomerAsync(Customer customer);
-        Task RemoveCustomerAsync(int cprNumber);
+        Task RemoveCustomerAsync(long cprNumber);
         Task<IList<Customer>> GetAllCustomersAsync();
         Task<String> CreateAccountAsync(string username);
     }
