+++ Group6System-main/Group6System-main/Database/DatabaseTier/Repository/AdminREPO/IAdminRepository.cs
@@ -9,7 +9,7 @@ namespace DatabaseTier.Repository.AdminREPO
     {
         Task<Customer> ValidateCustomerAsync(Customer customer);
         Task<IEnumerable<Customer>> GetAllCustomersAsync();
-        Task<string> RemoveCustomerAsync(int cprNumber);
+        Task<string> RemoveCustomerAsync(long cprNumber);
         Task<Account> CreateAccountAsync(Account account);
         Task<long> GetLastAccountNumberAsync();
         Task<User> CheckUserAsync(string username);
